This commit fixes two small bugs in nautilus: 1. the assignment operator on val pointers was not correctly called. For now, we defined it in the specialization. But we should add additional tests in nautilus. 2. the return op combination during the SSA assignment was not handling void returns correctly. This lead to an wrong argument propagation.
Furthermore, I translated some uint64_t to size_t as uint64_t is not equal to size_t on osx.

+++ nes-execution/src/Execution/Operators/Streaming/Aggregations/KeyedTimeWindow/KeyedWindowEmitAction.cpp
@@ -73,7 +73,7 @@ void KeyedWindowEmitAction::emitSlice(ExecutionContext& ctx,
         resultWindow.write(endTsFieldName, windowEnd);
         // load keys and write them to result record
         auto sliceKeys = globalEntry.getKeyPtr();
-        for (nautilus::static_val<uint64_t> i = 0; i < resultKeyFields.size(); ++i) {
+        for (nautilus::static_val<size_t> i = 0; i < resultKeyFields.size(); ++i) {
             const auto value = Nautilus::readExecDataTypeFromMemRef(sliceKeys, keyDataTypes[i]);
             resultWindow.write(resultKeyFields[i], value);
             sliceKeys = sliceKeys + nautilus::val<uint64_t>(keyDataTypes[i]->size());
